cast fix for 64 bit.

+++ PltApp3D.cpp
@@ -1,6 +1,6 @@
 
 //
-// $Id: PltApp3D.cpp,v 1.57 2004-11-15 22:16:38 vince Exp $
+// $Id: PltApp3D.cpp,v 1.58 2004-11-15 22:48:07 vince Exp $
 //
 
 // ---------------------------------------------------------------
@@ -923,7 +923,7 @@ void PltApp::DoOrient(Widget w, XtPointer client_data, XtPointer) {
   Real rW(1.0), rX(0.0), rY(0.0), rZ(0.0);
   Real rW2(1.0), rX2(0.0), rY2(0.0), rZ2(0.0);
   //int iOrientation((int) client_data);
-  int iOrientation(reinterpret_cast<int> (client_data));
+  long iOrientation(reinterpret_cast<long> (client_data));
   switch(iOrientation) {
     case OXY:
       rW  =  1.0;
